Rename ResponseStatus to CLIResponse::Status

+++ cliService/test/tests/cli/CommandHistoryTest.cpp
@@ -19,28 +19,28 @@ namespace cliService
   {
     EXPECT_TRUE(_history->getPreviousCommand().empty());
     EXPECT_TRUE(_history->getNextCommand().empty());
-    EXPECT_EQ(_history->size(), 0);
+    EXPECT_EQ(_history->size(), (size_t)0);
   }
 
   TEST_F(CommandHistoryTest, AddSingleCommand)
   {
     _history->addCommand("command1");
 
-    EXPECT_EQ(_history->size(), 1);
+    EXPECT_EQ(_history->size(), (size_t)1);
     EXPECT_EQ(_history->getPreviousCommand(), "command1");
   }
 
   TEST_F(CommandHistoryTest, IgnoreEmptyCommands)
   {
     _history->addCommand("");
-    EXPECT_EQ(_history->size(), 0);
+    EXPECT_EQ(_history->size(), (size_t)0);
   }
 
   TEST_F(CommandHistoryTest, IgnoreDuplicateCommands)
   {
     _history->addCommand("command1");
     _history->addCommand("command1");
-    EXPECT_EQ(_history->size(), 1);
+    EXPECT_EQ(_history->size(), (size_t)1);
   }
 
   TEST_F(CommandHistoryTest, RespectMaxSize)
@@ -50,7 +50,7 @@ namespace cliService
     _history->addCommand("command3");
     _history->addCommand("command4");
 
-    EXPECT_EQ(_history->size(), 3);
+    EXPECT_EQ(_history->size(), (size_t)3);
     EXPECT_EQ(_history->getPreviousCommand(), "command4");
     EXPECT_EQ(_history->getPreviousCommand(), "command3");
     EXPECT_EQ(_history->getPreviousCommand(), "command2");
@@ -87,7 +87,7 @@ namespace cliService
     _history->addCommand("command2");
 
     _history->clear();
-    EXPECT_EQ(_history->size(), 0);
+    EXPECT_EQ(_history->size(), (size_t)0);
     EXPECT_TRUE(_history->getPreviousCommand().empty());
     EXPECT_TRUE(_history->getNextCommand().empty());
   }
@@ -97,18 +97,18 @@ namespace cliService
     _history->addCommand("command1");
     _history->addCommand("command2");
 
-    EXPECT_EQ(_history->getCurrentIndex(), 2);
+    EXPECT_EQ(_history->getCurrentIndex(), (size_t)2);
     _history->getPreviousCommand();
-    EXPECT_EQ(_history->getCurrentIndex(), 1);
+    EXPECT_EQ(_history->getCurrentIndex(), (size_t)1);
     _history->getNextCommand();
-    EXPECT_EQ(_history->getCurrentIndex(), 2);
+    EXPECT_EQ(_history->getCurrentIndex(), (size_t)2);
   }
 
   TEST_F(CommandHistoryTest, DownArrowAsFirstAction)
   {
     // This test verifies the fix for the segmentation fault issue
     EXPECT_TRUE(_history->getNextCommand().empty());
-    EXPECT_EQ(_history->getCurrentIndex(), 0);
+    EXPECT_EQ(_history->getCurrentIndex(), (size_t)0);
   }
 
 }
